Started next part off project
Made ready to start next part off the project.
Made claude go trough and add some comment so if get back later it will
go faster but probaly a bit waste with how small the files is.
When started it felt how easy it is to work with later when add small c
project that is not worth it's own file but will try to keep files 1-9