Moved comments, fixed scanf in main.c
Moved comment outside the function so it's way cleaner and less noise.
Fixed so scanf in main.c is using the read functions in calculations.c
for it
Removed alot of uneeded inline comments.
Reduced noise thinking about making a readme.md or move docstring to .h
file to get it even less noise