Driving distance time
Created a new function that takes 2 input speed and distance and it
calculate the time you have travled or need to travel down from a hour
to a milisecond from there.
Added docstring for new function so it explains it good starting to hate
inline comment when started using docstring proparly.
Only run some test yet but seems to work as intended with the test ran.